% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_contain_reserved_col_name.R
\name{.contain_reserved_col_name}
\alias{.contain_reserved_col_name}
\title{Check if rowData(spe) contains reserved name}
\usage{
.contain_reserved_col_name(col_name)
}
\arguments{
\item{col_name}{the colnames}
}
\value{
TRUE when col_name contains reserved names, FALSE
}
\description{
Internal Funciton
}
